refactor: loadBitmapFromUri 함수 BitmapCropUtil object로 이동
