--- client/presentation/src/main/java/com/andone/memorip/presentation/util/BitmapFromUri.kt
@@ -1,18 +1,0 @@
-package com.andone.memorip.presentation.util
-
-import android.content.Context
-import android.graphics.Bitmap
-import android.graphics.BitmapFactory
-import android.net.Uri
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.withContext
-
-suspend fun loadBitmapFromUri(context: Context, uri: Uri): Bitmap? {
-    return withContext(Dispatchers.IO) {
-        runCatching {
-            context.contentResolver.openInputStream(uri)?.use { stream ->
-                BitmapFactory.decodeStream(stream)
-            }
-        }.getOrNull()
-    }
-}
